Trying to get parallel connections to work.  In the process of fixing a bug in which calls to cmm_select ignore new file descriptors that are connected as the result of a send that happens after select is called.
Solution: interrupt the select with a signal when new connections happen.
Problem: there's a bug in the 2.6.22.14 kernel that causes
pthread_cond_timedwait to return ETIMEDOUT prematurely if it's interrupted
by a signal.  With correct behavior, it should resume waiting on the
condition variable, and the caller shouldn't notice the interruption.
I found a patch for this bug, and will test it when the kernel is done
compiling.


git-svn-id: file:///home/svnshare/juggler_2.6.22.14/libcmm/trunk@334 1971e5c1-a347-0410-b49b-82492f0cf60c