In the middle of re-working sender-side and receiver-side data structures for pending IROBs, since they are actually not the same. Trying to work out what needs to be in both sender and receiver and what is unique to each.
Keep in mind that though the dependency-based ordering is enforced
at the receiver, the sender still needs dependency information in order
to make scheduling decisions and priority inversion.

Also, it would probably be a good idea to have the lattice structure
in the code, so that it's easy to get at the inter-IROB relationships.


git-svn-id: file:///home/svnshare/juggler_2.6.22.14/libcmm/trunk@376 1971e5c1-a347-0410-b49b-82492f0cf60c