Workaround for a weird issue probably caused by a misconfigured firewall. The issue as it manifests: 1) After a CSocket is idle for a while, I try to send something on it. 2) The CSocketReceiver dies on recv with ECONNRESET. 3) The CSocketSender notices, clears out the old connection, and attempts a new one. 4) The remote side already has a CSocket on that iface pair, so it ignores the connection.    -This was the bug.  It should have at least closed the new socket if it wanted to     disallow the duplicate connection, rather than just ignoring it. 5) The connecting side hangs waiting for the CSocket hello response.
Workaround: the accepting side trusts the connecting side, so it
ditches its old CSocket on that iface pair and accepts the new one.
This plus the misconfigured firewall results in some delays due to the
extra TCP handshakes, but it allows things to keep moving until I figure out
what's causing the resets.